requirements.txt
# We can not refer to an environment.yml file from another
# So to get performant launches on mybinder.org, we have copied
# the contents of this file to binder/environment.yml. If you
# make changes here, please copy them over there too.

# For Spyder core
applaunchservices >=0.1.7
atomicwrites >=1.2.0
chardet >=2.0.0
cloudpickle >=0.5.0
diff-match-patch >=20181111
intervaltree
IPython >=4.0
jedi ==0.15.2
keyring
nbconvert >=4.0
numpydoc >=0.6.0
Paramiko >=2.4.0
parso ==0.5.2
pexpect >=4.4.0
pickleshare >=0.4
psutil >=5.3
pycodestyle
pyflakes
pygments >=2.0
pylint >=0.25
pyqt5 <5.13
pyqtwebengine <5.13
python-language-server >=0.31.2,<0.32.0
pyxdg >=0.26
pyzmq >=17
qdarkstyle >=2.8
qtawesome >=0.5.7
qtconsole >=4.6.0
qtpy >=1.5.0
sphinx >=0.6.6
spyder-kernels >=1.9.0,<1.10.0
watchdog

# Incorporate packages for use by end user
numpy
scipy
pandas
matplotlib
cython
sympy

# For building standalone Mac app
py2app
dmgbuild
